requirements.txt
aiogram==3.3.0
SQLAlchemy==2.0.25
asyncpg==0.29.0
pydantic==1.10.13
python-dotenv==1.0.1
aiogram>=3.0.0

